backend/requirements.txt
fastapi==0.102.0
pydantic==2.3.0
email-validator==2.0.0
uvicorn==0.23.2
Jinja2==3.1.2
cassandra-driver==3.28.0
bcrypt==4.0.1